Revert the change to 'C100'
There are several different strings which will need to be replaced
once the name of the new service and its core data model have been
chosen. Rather than pre-empt that, I'm leaving the original values
in place. This will make it easier to go and search for those
values in the taxtribs repo, when the developer needs to
understand what they signify.