Bumping gems and sorting out form elements
Bumping some of our gems to the latest versions, and specifically
`govuk_elements_form_builder` which allow us to use the latest versions
of `govuk_elements_rails ` and `govuk_frontend_toolkit `.

Also removed Dropzone references, javascript and styles as it is not
going to be used for the time being.