Introduce the MIAM Acknowledgement page
This acts as a second start page after the entrypoint pages and it is
still not a step page as will be informative only too.
Will be after this page the application will efectively start.

Content as per current prototype, to have something to play with, but
will definetely change in the future.