Enable a basic CRUD for Applicants
Following the previous PR, this will enable the ability to add multiple
Applicants, and also edit and delete existing ones.

Some of the code will be candidate to extract to superclasses or modules
once we are happy with the solution and also once we have another similar
step (like Respondents) where we will be able to reuse some stuff.

There are however some rough corners in this solution, which hopefully we
will be polishing as part of following PRs.

A very ugly little interface has been incorporated to the top of the step
view to be able to manipulate the Applicant entries visually. Of course this
will change in the future.