Whitelist gov_uk_date field, if any
For the fields generated by the gem to be retrieved in the controller,
those need to be whitelisted. These fields are (in the case of `dob`):
dob_dd, dob_mm and dob_yyyy.

As there could be more than a gov_uk_date field per form, the method
will iterate through all the dates and build all the attribute names.