Cleanup humongous mass assignments
Simplified and removed duplication of attribute mass assignments in
personal details controllers and form objects.

This same approach can be extended to any other controller or form
object, but the ones we have at the moment are pretty simple with only
one or two attributes, so for now it is not worth changing those.