Add step `children names`
This is a step using most of the code already created for the concern
`CrudStep` but with some modifications in the form object to be able to
create several children records at once, or update existing ones.

Next children steps now need to be adapted to use the children records
created by this step, to match the prototype, as these are now outdated.
This will be done in a follow-up PR.