Speed up mutant, reducing scope by default
As the project grows, mutant is taking a huge amount of time to finish,
and lately Travis can't handle it, killing the build after 45 mins or so.

With these changes, by default mutant will run in a smaller sample of
classes but still can be run on the whole project manually.

Updated README to explain the options.