Extract exemption locales to separate file
As the exemption locales are going to grow a lot with following steps,
it makes sense to separate them to their own locale file. For now just
one file, as it is easy in the dev phase, but later on if it grows too
much we can split it into more files.

Thanks @tatyree for the suggestion.