Rework the `Answer` class
The ‘raw’ mechanism was a bit out of place for this project as those
free-text answers left blank still need to be printed to the PDF
(saying ‘Not applicable’ or similar wording).

Organised in a way we can inherit from a BaseAnswer presenter and to
have different kind of answers with different templates and be more
flexible.