Parametrise Answer `show` mechanism
Unfortunately, sometimes we need to show the question even if the
answer is nil/empty (in which case we print `Not applicable` or
similar). But other times if the answer is empty we don’t need to show
the question.

So… allow to pass an optional `show` value to cope with this.