Implement C8 details confidentiality
If the applicant enables the confidentiality of personal details,
then the C100 PDF form will hide a limited set of details and will
generate an extra C8 form with the omitted details.

This PR only implements the 'hiding' of details in a flexible way,
but not yet the output of the C8 form.