Refactor of the PDF generation to combine multiple forms
We have to generate the C100 x 3 times, but also if triggered, the
C1A x 3 times, and if triggered, the C8 x 1 time. Also, 1 cover letter.

Lots of documents, that need to be paginated 1/X correctly, and need to
support custom headers and footers.
Also for efficiency it is much better to generate the C100 once, and
copy it 3 times, (and the same with the C1A).

Individual documents can still be generated, so for example we can still
send by email 1 copy of the C100 (instead of the whole bundle with all
contained inside). If needed.