C8 handling for `other parties` relationships
If confidentiality has been triggered, then we need to completely
hide `other parties` relationships, if there are any other parties.

Also eventually we will need to do a similar thing for the residence
details, but leaving it for now until I have confirmation.

A bit of refactor to split into two separate section presenters the
relationships and the residence.