Create basic emails, and templates config
We had already a NotifyMailer from TaxTribs, but I’ve removed unneeded
code (at least for now) and improved the way we handle common
personalisation like the service name.

Also, get rid of ENV variables for template IDs, as they are not
secrets, we can declare them in a config file for more easy maintenance.