Implement the reminder mechanism
This is mostly the same solution we had in Tax Tribs, with minor changes.

We setup 2 draft expire reminders, one to be sent when 5 days left and another one to be sent when 1 day left.
The mechanism is flexible enough to setup as many reminders as we want to trigger at specific moments in time.