Basic `check your answers`
This is a first version of the check your answers page.

Only the first few questions (MIAM attendance) are shown, and in any case this is not yet integrated into the user journey so will only be visible if the URL is known.

We can iterate on this and if we are happy, start adding more and more question/answers.