Reorganise children sections in CYA
This follows more closely the pattern we used for the rest of people (applicants, respondents...) and also simplify the code.

The relationships has been removed from the children sections, and will be reintroduced in each of the applicant/respondent sections, as structurally, it is were they belong when the user is completing their application online.