Modify resume draft to use CYA
Again, this is behind a feature flag so for now, no real users will be able to access it.

When resuming a draft, now instead of taking you directly to the last step, there will be a middle page which is the `Check your answers`, to check the current progress and from there to be able to resume (will take you to the last completed step).