Add a reference code to PDF and confirmation page
This reference code is just a simple short version of the created_at, and the UUID of the application.

The purpose is mainly give the user peace of mind having a reference code to confirm their application has been sent and to be able to cross-reference in the future PDF with their codes in case the user wants to know about their application.

Will also be sent in the receipt email if the user chose to.