Very basic application stats (#444)
* Very basic applications stats

As our retention policy is quite restrictive (28 days for applications), and in order to quickly answer question like `how many applications were submitted last week`, etc. we need a way to persist totals for the most common metrics.

In GA we have almost all of this information too, so not really spending too much time here duplicating something GA do quite well.

Run daily through `daily_tasks.rake`

* Adding more data points

And a bit of refactor to make the dates more flexible (so we can fill a months worth of data once we release it).