Add a simple URL shortener feature (#446)
This works together with our new 'shorter' domain CNAME `c100.dsd.io`

Short urls can be added and/or listed via a rake task as a first step. In the future, and if needed, we can expose a simple CRUD admin page, but at the moment this is the MVP to be able to produce short URLs with tracking.