Undo the `CumulativeData` stats
It turns out this is not what we needed. We just need to track the number of applications per court, for longer than 28 days.

In order to do that, we will take a different approach and create a "kind of" audit table to track that specific information, in a separate PR.