Restore target attribute in external links
Mobile devices seem to handle properly the opening of these external links in a new tab, but desktop browsers (at least Safari and Chrome) are having trouble.

This has something to do with the JS code we are using to intercept these links and track hits to GA.

Adding back the target (it was there some time ago) to be a explicit link attribute in order to err on the side of caution so desktop browsers open new tabs.