Fix bug involving OtherChild residence
There was a bug that only now has made appearance due to a Sentry error trace, where we were using the 'minors' collection instead of the 'children' collection in the Residence step.

https://sentry.service.dsd.io/mojds/c100-application-prod/issues/33305/

It was working fine because when using `minor_ids.first` it was always returning a `Child` type first, instead of an `OtherChild`, but, as an edge case, if the applicant deleted some or all of the Child and added later some OtherChild, will alter the returned elements.

This has now been corrected by using always the proper collection (child_ids) for residence, so regardless the order, it can't return an `OtherChild`.