Send application with Notify (feature-flag)
Add a new mailer, inheriting from NotifyMailer, to handle the 2 emails that we were sending with Amazon SES, making use of Notify document upload functionality.

For now this is behind a feature flag, so only "visible" on local envs and staging (`DEV_TOOLS_ENABLED`).

Once we test it and we are happy with it, we will clean up the code and replace Amazon SES with these.