Tune a bit puma killer
Puma killer is quite pessimistic with the amount of available free memory, so it would kill the workers when there is still enough memory.

We can use the env variable to set the available ram slighly higher (about 120mb higher) than the actual ram, so puma killer do a more accurate job.