Update multiple names UI
We now ask for both first name and last name in separate text fields
rather than one.

This could potentially lead to confusion when there are multiple names
as there isn't as good distinction between groups of names.

This change switches the font weight to create a visual separation
between the form groups.