Make `remove name` work without JS
We had a functionality depending on JS being enabled. This is not ideal and we've refactor the code to get rid of the JS dependency and as a side effect enable the possibility of remove individual names, no matter their position in the list.

As a consequence of this change, there is now some styling TLC and possibly the need to disable the ENTER key with JS, which is a minor inconvenience.