Update CircleCI workflow to promote to production
We've created the k8s production namespace and now we can start promoting builds to k8s production.

The worflow have been updated to support this.

Also, removed a conditional in the `build_staging` job, checking if we were in master branch, because we always are in master if this job runs (there is filter set to 'only master' in the workflow).