Add clearer label to date of birth unknown checkbox
The date of birth unknown checkbox isn't inside of the fieldset for date of birth. This means that a screen reader user wouldn't be given the context of 'date of birth' when tabbing onto the element. This label clarifies what it is about, so that the legend isn't neccessary, as adding it into the fieldset would be a lot more work.