Implement a BareApplicationController superclass
There is some code shared among the Status controller and the Callbacks controller.

In order to DRY a bit this, we move this shared code to a superclass that will be used in those controllers.

Make sure we log the referrer and user_agent as these were not included before in any of the requests to these controllers.

Also, get rid of the brakeman exceptions by actually calling `protect_from_forgery` because, for GET requests, it will do nothing anyway.