Add a forgery protection test in callbacks controller
It was too easy to break the `#notify` callback endpoint (which receives a POST) just by messing around with the `protect_from_forgery` config.

Ensure we cover this scenario in case it happens again in the future.
This endpoint must not have forgery protection enabled because we are using instead a Bearer token for a custom authentication.