Refactor and reword error messages - Part 1
We need to reword most of the error messages in the application forms, in order to comply with accessibility and design guidelines.

As this is going to take a while and a lot of changes in the locales are expected, let's create smaller PRs as we go, so there is less risk of introducing regressions.

This first PR will create a new `locales/errors/en.yml` file where we move all the form errors previously in the general `en.yml`, and start adding/rewording some of the errors.