Refactor the online submission jobs
Follow up to PR #638.

This is just a refactor with not breaking changes. Current behaviour is maintained.

As we intend to split the C8 form out of the PDF bundle, in the email to court, but not in the email to the applicant, we can't continue using a common, shared, `OnlineSubmission` class because it will not allow us to generate different combinations of documents (without having to rely on too many if-elses).

The cleanest and simple solution, is to split the `OnlineSubmission` into two specialised classes, one for the Applicant and another one for the Court.

As we now have two specialised classes we can also bring back all the code that previously we had in `PdfEmailSubmission` as now is better encapsulated in their individual submission classes and let us get rid of some ugly if-elses and parametrisation (`recipient: :applicant` /  `recipient: :court`).

In following PRs we will introduce the generation of different documents for the email to court. This is why in this PR there is a bit of code duplication in the `#generate_pdf` method, and will be removed in the next PR.