Add new template for C8 split
We must maintain current behaviour on production, but still be able to test the new behaviour on staging/local.

Unfortunately with Notify this is tricky, so temporary we create a new template and use one or another based on the feature flag.

Once we decide to enable this for all users in production, we can get rid of the new template and just modify the current existing one. And also cleanup the if-else and other ugly code.