Make the lookup conditional based on feature-flag
During the development phase we want this to be enabled only on developer machines based on a feature-flag.

It is still not solid or complete enough to be released to staging.

Also note, as this feature requires the address to be split, the version must be at least 3 (same as split address feature).