Parameterize number of concurrent mutant jobs
When running on CI, a very high number of concurrent jobs seem to make some mutations fail randomly. Reducing the number of jobs minimise this problem.

We will reduce the number of jobs from the current 36 to 24 (but we might still need to tweak this number a bit).

On local machines the number of jobs will not be overriden, as per current behaviour.