Add convenience methods to Address struct
We will need this methods later on to show a shorter version of the address in the dropdown, for the user to select from.

Also all the details of the selected address need to be tokenised as a simple string, as html select options can't have anything other than strings as its values.

Finally, a custom attribute `SplitAddress` will take care of coercing the tokenised string back into a hash so we can update the DB record on `#persist`