Reset address lookup dependent attributes
This is a kind of edge case but... if the user search for a postcode and select an address from the list, but then they go back (for whatever reason) and search another postcode, and, without selecting a new address (could happen the lookup didn't return results) go to the address edit page, then the old address is still pre-populated, which is confusing for the user.

We avoid this by checking if the postcode has changed or not, to reset dependent attributes.

Also, we pre-populate the country as we know 99% of users will be in United Kingdom if using the postcode lookup, although for the cases where not (Isle of Man or Channel Islands) they can always change it.
This was a suggestion by Simon to improve the UX for users that couldn't find their address in the list.