Ensure a predictable full address order
The order in the `address_data` Hash (store) is not guaranteed, and in fact, some keys (postcode or country) are created before others (address lines).

This has the side effect of sometimes returning a funny `full_address` that does not maintain the order people are used to.

Make the `#full_address` method resilient to these scenarios and always return the attribute values in the same order they were declared, regardless of the order of the keys in the Hash.