Add postcode lookup errors to sentry (#668)
* add postcode lookup errors to sentry

* Refactor error handling and specs

We are going to get rid of the `AddressLookupErrors` because it is not used in the end. We really just need a way to know if there was an error, or not, so we are going to use a simple `attr_reader :last_exception` in the `AddressLookupService` class to store the last exception, if any, otherwise will be nil.

Also, in case of unsuccessful response from the Places API, we will raise a custom `UnsuccessfulLookupError`.

This simplify the tests, and mutant is also more happy now.