Enable the address lookup in production
We have the green light. Up until now this feature has been only visible on staging and local dev environments.

Still, we must ensure this is only enabled for `version >= 3` of the applications, so the condition will remain until all applications are in the new version, which can still take a couple weeks.