Protect from nil values
Under certain circunstances, an applicant might leave out some mandatory follow-up questions after changing their answer to the `abduction risk` or the `previous proceedings` questions.

The result of this would be an exception when trying to generate the PDF, due to missing data.

Although this has occurred only twice in the past so far, we better protect from this from happening again.