Force garbage collection on delivery jobs
We continue to have memory peaks after each submission. This is mostly a consequence of the PDF generation, and the tools used to combine multiple files into one.

Let's experiment for a few days by having an explicit garbage collection after the jobs complete, and see if this  improves in any way the situation.

As these are async jobs, there is no penalty to the user (that otherwise would be if this was part of the request lifecycle).