Run the sanity check validation
On the step that actually starts the application (marks it as `in progress`) we are going to do an extra validation now, to ensure the screener has been completed.

Before this, there was an edge case where once you obtain a session (just visiting the postcode step) and then jumping manually to the `child_protection_cases` step, would actually let you pass with a corrupt or missing 'screening', thus failing at submission time.