Increase date range in DB lookup
There was an edge case where applications created in the last day of the month were not found when using the utility method `#find_by_reference_code`.

To ensure these applications are also included, we will force the end of the day, so the where clause is (take May as an example):

`WHERE (created_at BETWEEN '2019-05-01 00:00:00' AND '2019-05-31 23:59:59.999999')`

This is an internal method for debug purposes and is not used anywhere, so no harm done, but to avoid confusion in the future better to change it.