Enable postgres `verify-full` SSL mode
For production environments we are going to force the `verify-full` SSL mode in all connections to the database.

It will be overriden (`DATABASE_SSLMODE=disable`) when using docker-compose so we are still able to run and test the application locally.

This follows best practices and guidance from Cloud Platforms here:

https://user-guide.cloud-platform.service.justice.gov.uk/tasks.html#ssl-connections-with-rds

To avoid potential confusion with the introduction of this new ENV var, I'm renaming `DISABLE_SSL` to `DISABLE_HTTPS` so it is more clear this other env variable is for https connections and nothing to do with the database.